AddSamples {Seurat}	R Documentation

Add samples into existing Seurat object.

Description

Add samples into existing Seurat object.

Usage

AddSamples(object, new.data, min.genes = 2500, names.field = 1,
  names.delim = "_")

Arguments

object
Seurat object

new.data
Data to be added to the object

min.genes
Gene threshold for new data. Discard all cells with fewer genes detected

names.field
For the initial identity class for each cell, choose this field from the cell's column name

names.delim
For the initial identity class for each cell, choose this delimiter from the cell's column name

[Package Seurat version 1.3.5 Index]