AverageExpression {Seurat}	R Documentation

Averaged gene expression by identity class

Description

Returns gene expression for an 'average' single cell in each identity class

Usage

AverageExpression(object, genes.use = NULL, return.seurat = F,
  add.ident = NULL, ...)

Arguments

object
Seurat object

genes.use
Genes to analyze. Default is all genes.

return.seurat
Whether to return the data as a Seurat object. Default is false.

add.ident
Add identity label to class

...
Arguments to be passed to methods such as Setup

Details

Output is in log-space, but averaging is done in non-log space.

Value

Returns a matrix with genes as rows, identity classes as columns.

[Package Seurat version 1.3.5 Index]