BatchGene {Seurat}	R Documentation

Identify potential genes associated with batch effects

Description

Test for genes whose expression value is strongly predictive of batch (based on ROC classification). Important note: Assumes that the 'batch' of each cell is assigned to the cell's identity class (will be improved in a future release)

Usage

BatchGene(object, idents.use, genes.use = NULL, auc.cutoff = 0.6,
  thresh.use = 0)

Arguments

object
Seurat object

idents.use
Batch names to test

genes.use
Gene list to test

auc.cutoff
Minimum AUC needed to qualify as a 'batch gene'

thresh.use
Limit testing to genes which show, on average, at least X-fold difference (log-scale) in any one batch

Value

Returns a list of genes that are strongly correlated with batch.

[Package Seurat version 1.3.5 Index]