ClusterAlpha {Seurat}	R Documentation

Probability of detection by identity class

Description

For each gene, calculates the probability of detection for each identity class.

Usage

ClusterAlpha(object, thresh.min = 0)

Arguments

object
Seurat object

thresh.min
Minimum threshold to define 'detected' (log-scale)

Value

Returns a matrix with genes as rows, identity classes as columns.

[Package Seurat version 1.3.5 Index]