DBClustDimension {Seurat}	R Documentation

Perform spectral density clustering on single cells

Description

Find point clounds single cells in a two-dimensional space using density clustering (DBSCAN).

Usage

DBClustDimension(object, dim.1 = 1, dim.2 = 2, reduction.use = "tsne",
  G.use = NULL, set.ident = TRUE, seed.use = 1, ...)

Arguments

object
Seurat object

dim.1
First dimension to use

dim.2
second dimension to use

reduction.use
Which dimensional reduction to use (either 'pca' or 'ica')

G.use
Parameter for the density clustering. Lower value to get more fine-scale clustering

set.ident
TRUE by default. Set identity class to the results of the density clustering. Unassigned cells (cells that cannot be assigned a cluster) are placed in cluster 1, if there are any.

seed.use
Random seed for the dbscan function

...
Additional arguments to be passed to the dbscan function

[Package Seurat version 1.3.5 Index]