DiffExpTest {Seurat}	R Documentation

Likelihood ratio test for zero-inflated data

Description

Identifies differentially expressed genes between two groups of cells using the LRT model proposed in Mcdavid et al, Bioinformatics, 2011

Usage

DiffExpTest(object, cells.1, cells.2, genes.use = NULL, thresh.use = log(2),
  print.bar = TRUE)

Arguments

object
Seurat object

cells.1
Group 1 cells

cells.2
Group 2 cells

genes.use
Genes to test. Default is to use all genes.

thresh.use
Limit testing to genes which show, on average, at least X-fold difference (log-scale) between the two groups of cells. Default is 0.25 Increasing thresh.use speeds up the function, but can miss weaker signals.

print.bar
Print a progress bar once expression testing begins (uses pbapply to do this)

Value

Returns a p-value ranked matrix of putative differentially expressed genes.

[Package Seurat version 1.3.5 Index]