DimPlot {Seurat}	R Documentation

Dimensional reduction plot

Description

Graphs the output of a dimensional reduction technique (PCA by default). Cells are colored by their identity class.

Usage

DimPlot(object, reduction.use = "pca", dim.1 = 1, dim.2 = 2,
  cells.use = NULL, pt.size = 3, do.return = FALSE, do.bare = FALSE,
  cols.use = NULL, group.by = "ident", pt.shape = NULL,
  do.label = FALSE, label.size = 1, no.legend = FALSE)

Arguments

object
Seurat object

reduction.use
Which dimensionality reduction to use. Default is "pca", can also be "tsne", or "ica", assuming these are precomputed.

dim.1
Dimension for x-axis (default 1)

dim.2
Dimension for y-axis (default 2)

cells.use
Vector of cells to plot (default is all cells)

pt.size
Adjust point size for plotting

do.return
Return a ggplot2 object (default : FALSE)

do.bare
Do only minimal formatting (default : FALSE)

cols.use
Vector of colors, each color corresponds to an identity class. By default, ggplot assigns colors.

group.by
Group (color) cells in different ways (for example, orig.ident)

pt.shape
If NULL, all points are circles (default). You can specify any cell attribute (that can be pulled with FetchData) allowing for both different colors and different shapes on cells.

do.label
Whether to label the clusters

label.size
Sets size of labels

no.legend
Setting to TRUE will remove the legend

Value

If do.return==TRUE, returns a ggplot2 object. Otherwise, only graphical output.

[Package Seurat version 1.3.5 Index]