DotPlot {Seurat}	R Documentation

Dot plot visualization

Description

Intuitive way of visualizing how gene expression changes across different identity classes (clusters). The size of the dot encodes the percentage of cells within a class, while the color encodes the AverageExpression level of 'expressing' cells (green is high).

Usage

DotPlot(object, genes.plot, cex.use = 2, cols.use = NULL,
  thresh.col = 2.5, dot.min = 0.05, group.by = NULL, ...)

Arguments

object
Seurat object

genes.plot
Input vector of genes

cex.use
Scaling factor for the dots (scales all dot sizes)

cols.use
Colors to use for plotting

thresh.col
The raw data value which corresponds to a red dot (lowest expression)

dot.min
The fraction of cells at which to draw the smallest dot (default is 0.05)

group.by
Group (color) cells in different ways (for example, orig.ident)

...
additional parameters to pass to FetchData (for example, use.imputed, use.scaled, use.raw)

Value

Only graphical output

[Package Seurat version 1.3.5 Index]