FeatureHeatmap {Seurat}	R Documentation

Vizualization of multiple features

Description

Similar to FeaturePlot, however, also splits the plot by visualizing each identity class separately.

Usage

FeatureHeatmap(object, features.plot, dim.1 = 1, dim.2 = 2,
  idents.use = NULL, pt.size = 2, cols.use = rev(heat.colors(10)),
  pch.use = 16, reduction.use = "tsne")

Arguments

object
Seurat object

features.plot
Vector of features to plot

dim.1
Dimension for x-axis (default 1)

dim.2
Dimension for y-axis (default 2)

idents.use
Which identity classes to display (default is all identity classes)

pt.size
Adjust point size for plotting

cols.use
Ordered vector of colors to use for plotting. Default is heat.colors(10).

pch.use
Pch for plotting

reduction.use
Which dimensionality reduction to use. Default is "tsne", can also be "pca", or "ica", assuming these are precomputed.

Details

Particularly useful for seeing if the same groups of cells co-exhibit a common feature (i.e. co-express a gene), even within an identity class. Best understood by example.

Value

No return value, only a graphical output

[Package Seurat version 1.3.5 Index]