FeaturePlot {Seurat}	R Documentation

Visualize 'features' on a dimensional reduction plot

Description

Colors single cells on a dimensional reduction plot according to a 'feature' (i.e. gene expression, PC scores, number of genes detected, etc.)

Usage

FeaturePlot(object, features.plot, dim.1 = 1, dim.2 = 2, cells.use = NULL,
  pt.size = 1, cols.use = c("yellow", "red"), pch.use = 16,
  reduction.use = "tsne", use.imputed = FALSE, nCol = NULL,
  no.axes = FALSE)

Arguments

object
Seurat object

features.plot
Vector of features to plot

dim.1
Dimension for x-axis (default 1)

dim.2
Dimension for y-axis (default 2)

cells.use
Vector of cells to plot (default is all cells)

pt.size
Adjust point size for plotting

cols.use
The two colors to form the gradient over. Provide as string vector with the first color corresponding to low values, the second to high. Also accepts a Brewer color scale or vector of colors. Note: this will bin the data into number of colors provided.

pch.use
Pch for plotting

reduction.use
Which dimensionality reduction to use. Default is "tsne", can also be "pca", or "ica", assuming these are precomputed.

use.imputed
Use imputed values for gene expression (default is FALSE)

nCol
Number of columns to use when plotting multiple features.

no.axes
Remove axis labels

Value

No return value, only a graphical output

[Package Seurat version 1.3.5 Index]