FetchData {Seurat}	R Documentation

Access cellular data

Description

Retreives data (gene expression, PCA scores, etc, metrics, etc.) for a set of cells in a Seurat object

Usage

FetchData(object, vars.all = NULL, cells.use = NULL, use.imputed = FALSE,
  use.scaled = FALSE, use.raw = FALSE)

Arguments

object
Seurat object

vars.all
List of all variables to fetch

cells.use
Cells to collect data for (default is all cells)

use.imputed
For gene expression, use imputed values

use.scaled
For gene expression, use scaled values

use.raw
For gene expression, use raw values

Value

A data frame with cells as rows and cellular data as columns

[Package Seurat version 1.3.5 Index]