FindAllMarkers {Seurat}	R Documentation

Gene expression markers for all identity classes

Description

Finds markers (differentially expressed genes) for each of the identity classes in a dataset

Usage

FindAllMarkers(object, ident.1, ident.2 = NULL, genes.use = NULL,
  thresh.use = 0.25, test.use = "bimod", min.pct = 0.1,
  print.bar = TRUE, only.pos = FALSE, max.cells.per.ident = Inf,
  return.thresh = 0.01, do.print = FALSE)

Arguments

object
Seurat object

ident.1
Identity class to define markers for

ident.2
A second identity class for comparison. If NULL (default) - use all other cells for comparison.

genes.use
Genes to test. Default is to use all genes.

thresh.use
Limit testing to genes which show, on average, at least X-fold difference (log-scale) between the two groups of cells. Increasing thresh.use speeds up the function, but can miss weaker signals.

test.use
Denotes which test to use. Seurat currently implements "bimod" (likelihood-ratio test for single cell gene expression, McDavid et al., Bioinformatics, 2011, default), "roc" (standard AUC classifier), "t" (Students t-test), and "tobit" (Tobit-test for differential gene expression, as in Trapnell et al., Nature Biotech, 2014)

min.pct
- only test genes that are detected in a minimum fraction of min.pct cells in either of the two populations. Meant to speed up the function by not testing genes that are very infrequently expression

print.bar
Print a progress bar once expression testing begins (uses pbapply to do this)

only.pos
Only return positive markers (FALSE by default)

max.cells.per.ident
Down sample each identity class to a max number. Default is no downsampling.

return.thresh
Only return markers that have a p-value < return.thresh, or a power > return.thresh (if the test is ROC)

do.print
FALSE by default. If TRUE, outputs updates on progress.

Value

Matrix containing a ranked list of putative markers, and associated statistics (p-values, ROC score, etc.)

[Package Seurat version 1.3.5 Index]