FindMarkersNode {Seurat}	R Documentation

Gene expression markers of identity classes defined by a phylogenetic clade

Description

Finds markers (differentially expressed genes) based on a branching point (node) in the phylogenetic tree. Markers that define clusters in the left branch are positive markers. Markers that define the right branch are negative markers.

Usage

FindMarkersNode(object, node, genes.use = NULL, thresh.use = log(2),
  test.use = "bimod", ...)

Arguments

object
Seurat object

node
The node in the phylogenetic tree to use as a branch point

genes.use
Genes to test. Default is to use all genes.

thresh.use
Limit testing to genes which show, on average, at least X-fold difference (log-scale) between the two groups of cells. Default is 0.25 Increasing thresh.use speeds up the function, but can miss weaker signals.

test.use
Denotes which test to use. Seurat currently implements "bimod" (likelihood-ratio test for single cell gene expression, McDavid et al., Bioinformatics, 2011, default), "roc" (standard AUC classifier), "t" (Students t-test), and "tobit" (Tobit-test for differential gene expression, as in Trapnell et al., Nature Biotech, 2014)

...
Additional arguments passed to FindMarkers

Value

Matrix containing a ranked list of putative markers, and associated statistics (p-values, ROC score, etc.)

[Package Seurat version 1.3.5 Index]