GetCentroids {Seurat}	R Documentation

Get cell centroids

Description

Calculate the spatial mapping centroids for each cell, based on previously calculated mapping probabilities for each bin.

Usage

GetCentroids(object, cells.use = NULL, get.exact = TRUE)

Arguments

object
Seurat object

cells.use
Cells to calculate centroids for (default is all cells)

get.exact
Get exact centroid (Default is TRUE). If FALSE, identify the single closest bin.

Details

Currently, Seurat assumes that the tissue of interest has an 8x8 bin structure. This will be broadened in a future release.

Value

Data frame containing the x and y coordinates for each cell centroid.

[Package Seurat version 1.3.5 Index]