ICAPlot {Seurat}	R Documentation

Plot ICA map

Description

Graphs the output of a ICA analysis Cells are colored by their identity class.

Usage

ICAPlot(object, ...)

Arguments

object
Seurat object

...
Additional parameters to DimPlot, for example, which dimensions to plot.

Details

This function is a wrapper for DimPlot. See ?DimPlot for a full list of possible arguments which can be passed in here.

[Package Seurat version 1.3.5 Index]