InitialMapping {Seurat}	R Documentation

Infer spatial origins for single cells

Description

Probabilistically maps single cells based on (imputed) gene expression estimates, a set of mixture models, and an in situ spatial reference map.

Usage

InitialMapping(object, cells.use = NULL)

Arguments

object
Seurat object

cells.use
Which cells to map

Value

Seurat object, where mapping probabilities for each bin are stored in object@final.prob

[Package Seurat version 1.3.5 Index]