LogNormalize {Seurat}	R Documentation

Normalize raw data

Description

Normalize count data per cell and transform to log scale

Usage

LogNormalize(data, scale.factor = 10000)

Arguments

data
Matrix with the raw count data

scale.factor
Scale the data. Default is 1e4

Value

Returns a matrix with the normalize and log transformed data

[Package Seurat version 1.3.5 Index]