MarkerTest {Seurat}	R Documentation

ROC-based marker discovery

Description

Identifies 'markers' of gene expression using ROC analysis. For each gene, evaluates (using AUC) a classifier built on that gene alone, to classify between two groups of cells.

Usage

MarkerTest(object, cells.1, cells.2, genes.use = NULL, thresh.use = log(2),
  print.bar = TRUE)

Arguments

object
Seurat object

cells.1
Group 1 cells

cells.2
Group 2 cells

genes.use
Genes to test. Default is to use all genes.

thresh.use
Limit testing to genes which show, on average, at least X-fold difference (log-scale) between the two groups of cells. Default is 0.25 Increasing thresh.use speeds up the function, but can miss weaker signals.

print.bar
Print a progress bar once expression testing begins (uses pbapply to do this)

Details

An AUC value of 1 means that expression values for this gene alone can perfectly classify the two groupings (i.e. Each of the cells in cells.1 exhibit a higher level than each of the cells in cells.2). An AUC value of 0 also means there is perfect classification, but in the other direction. A value of 0.5 implies that the gene has no predictive power to classify the two groups.

Value

Returns a 'predictive power' (abs(AUC-0.5)) ranked matrix of putative differentially expressed genes.

[Package Seurat version 1.3.5 Index]