PCASigGenes {Seurat}	R Documentation

Significant genes from a PCA

Description

Returns a set of genes, based on the JackStraw analysis, that have statistically significant associations with a set of PCs.

Usage

PCASigGenes(object, pcs.use, pval.cut = 0.1, use.full = TRUE,
  max.per.pc = NULL)

Arguments

object
Seurat object

pcs.use
PCS to use.

pval.cut
P-value cutoff

use.full
Use the full list of genes (from the projected PCA). Assumes that ProjectPCA has been run. Default is TRUE

max.per.pc
Maximum number of genes to return per PC. Used to avoid genes from one PC dominating the entire analysis.

Value

A vector of genes whose p-values are statistically significant for at least one of the given PCs.

[Package Seurat version 1.3.5 Index]