PCTopCells {Seurat}	R Documentation

Find cells with highest PCA scores

Description

Return a list of genes with the strongest contribution to a set of principal components

Usage

PCTopCells(object, pc.use = 1, num.cells = NULL, do.balanced = FALSE)

Arguments

object
Seurat object

pc.use
Principal component to use

num.cells
Number of cells to return

do.balanced
Return an equal number of cells with both + and - PC scores.

Value

Returns a vector of cells

[Package Seurat version 1.3.5 Index]