Read10X {Seurat}	R Documentation

Load in data from 10X

Description

Enables easy loading of sparse data matrices provided by 10X genomics.

Usage

Read10X(data.dir = NULL)

Arguments

data.dir
Directory containing the matrix.mtx, genes.tsv, and barcodes.tsv files provided by 10X.

Value

Returns a sparse matrix with rows and columns labeled

[Package Seurat version 1.3.5 Index]