RefinedMapping {Seurat}	R Documentation

Quantitative refinement of spatial inferences

Description

Refines the initial mapping with more complex models that allow gene expression to vary quantitatively across bins (instead of 'on' or 'off'), and that also considers the covariance structure between genes.

Usage

RefinedMapping(object, genes.use)

Arguments

object
Seurat object

genes.use
Genes to use to drive the refinement procedure.

Details

Full details given in spatial mapping manuscript.

Value

Seurat object, where mapping probabilities for each bin are stored in object@final.prob

[Package Seurat version 1.3.5 Index]