ReorderIdent {Seurat}	R Documentation

Reorder identity classes

Description

Re-assigns the identity classes according to the average expression of a particular feature (i.e, gene expression, or PC score) Very useful after clustering, to re-order cells, for example, based on PC scores

Usage

ReorderIdent(object, feature = "PC1", rev = FALSE, aggregate.fxn = mean,
  reorder.numeric = FALSE, ...)

Arguments

object
Seurat object

feature
Feature to reorder on. Default is PC1

rev
Reverse ordering (default is FALSE)

aggregate.fxn
Function to evaluate each identity class based on (default is mean)

reorder.numeric
Rename all identity classes to be increasing numbers starting from 1 (default is FALSE)

...
additional arguemnts (i.e. use.imputed=TRUE)

Value

A seurat object where the identity have been re-oredered based on the average.

[Package Seurat version 1.3.5 Index]