RunDiffusion {Seurat}	R Documentation

Run t-distributed Stochastic Neighbor Embedding

Description

Run t-SNE dimensionality reduction on selected features. Has the option of running in a reduced dimensional space (i.e. spectral tSNE, recommended), or running based on a set of genes

Usage

RunDiffusion(object, cells.use = NULL, dims.use = 1:5, k.seed = 1,
  do.fast = FALSE, add.iter = 0, genes.use = NULL,
  reduction.use = "pca", dim_embed = 2, q.use = 0.05, max.dim = 2,
  scale.clip = 3, ...)

Arguments

object
Seurat object

cells.use
Which cells to analyze (default, all cells)

dims.use
Which dimensions to use as input features

k.seed
Random seed for the t-SNE

do.fast
If TRUE, uses the Barnes-hut implementation, which runs faster, but is less flexible

add.iter
If an existing tSNE has already been computed, uses the current tSNE to seed the algorithm and then adds additional iterations on top of this

genes.use
If set, run the tSNE on this subset of genes (instead of running on a set of reduced dimensions). Not set (NULL) by default

reduction.use
Which dimensional reduction (PCA or ICA) to use for the tSNE. Default is PCA

dim_embed
The dimensional space of the resulting tSNE embedding (default is 2). For example, set to 3 for a 3d tSNE

q.use
Quantile to use

max.dim
Max dimension to keep from diffusion calculation

scale.clip
Max/min value for scaled data. Default is 3

...
Additional arguments to the tSNE call. Most commonly used is perplexity (expected number of neighbors default is 30)

Value

Returns a Seurat object with a tSNE embedding in object@tsne_rot

[Package Seurat version 1.3.5 Index]