ScaleData {Seurat}	R Documentation

Scale and center the data

Description

Scale and center the data

Usage

ScaleData(object, genes.use = NULL, data.use = NULL, do.scale = TRUE,
  do.center = TRUE, scale.max = 10)

Arguments

object
Seurat object

genes.use
Vector of gene names to scale/center. Default is all genes in object@data.

data.use
Can optionally pass a matrix of data to scale, default is object@data[genes.use,]

do.scale
Whether to scale the data.

do.center
Whether to center the data.

scale.max
Max value to accept for scaled data. The default is 10. Setting this can help reduce the effects of genes that are only expressed in a very small number of cells.

Value

Returns a seurat object with object@scale.data updated with scaled and/or centered data.

[Package Seurat version 1.3.5 Index]