SetIdent {Seurat}	R Documentation

Set identity class information

Description

Sets the identity class value for a subset (or all) cells

Usage

SetIdent(object, cells.use = NULL, ident.use = NULL)

Arguments

object
Seurat object

cells.use
Vector of cells to set identity class info for (default is all cells)

ident.use
Vector of identity class values to assign (character vector)

Value

A Seurat object where object@ident has been appropriately modified

[Package Seurat version 1.3.5 Index]