Seurat-deprecated {Seurat}	R Documentation

Deprecated function(s) in the Seurat package

Description

These functions are provided for compatibility with older version of the Seurat package. They may eventually be completely removed.

Usage

setup(...)

Arguments

...
Parameters to be passed to the modern version of the function

Details

setup 	 now a synonym for Setup

vlnPlot 	 now a synonym for VlnPlot

subsetData 	 now a synonym for SubsetData

mean.var.plot 	 now a synonym for MeanVarPlot

pca 	 now a synonym for PCA

project.pca 	 now a synonym for ProjectPCA

print.pca 	 now a synonym for PrintPCA

viz.pca 	 now a synonym for VizPCA

set.ident 	 now a synonym for SetIdent

pca.plot 	 now a synonym for PCAPlot

pcHeatmap 	 now a synonym for PCHeatmap

jackStraw 	 now a synonym for JackStraw

jackStrawPlot 	 now a synonym for JackStrawPlot

run_tsne 	 now a synonym for RunTSNE

tsne.plot 	 now a synonym for TSNEPlot

find.markers 	 now a synonym for FindMarkers

find_all_markers 	 now a synonym for FindAllMarkers

genePlot 	 now a synonym for GenePlot

feature.plot 	 now a synonym for FeaturePlot

buildClusterTree 	 now a synonym for BuildClusterTree

plotClusterTree 	 now a synonym for PlotClusterTree

plotNoiseModel 	 now a synonym for PlotNoiseModel

add_samples 	 now a synonym for AddSamples

subsetCells 	 now a synonym for SubsetCells

project.samples 	 now a synonym for ProjectSamples

run_diffusion 	 now a synonym for RunDiffusion

ica 	 now a synonym for ICA

cluster.alpha 	 now a synonym for ClusterAlpha

reorder.ident 	 now a synonym for ReorderIdent

average.pca 	 now a synonym for AveragePCA

average.expression 	 now a synonym for AverageExpression

icTopGenes 	 now a synonym for ICTopGenes

pcTopGenes 	 now a synonym for PCTopGenes

pcTopCells 	 now a synonym for PCTopCells

fetch.data 	 now a synonym for FetchData

viz.ica 	 now a synonym for VizIca

getWeightMatrix 	 now a synonym for GetWeightMatrix

regulatorScore 	 now a synonym for RegulatorScore

find.markers.node 	 now a synonym for FindMarkersNode

diffExp.test 	 now a synonym for DiffExpTest

tobit.test 	 now a synonym for TobitTest

batch.gene 	 now a synonym for BatchGene

marker.test 	 now a synonym for MarkerTest

diff.t.test 	 now a synonym for DiffTTest

which.cells 	 now a synonym for WhichCells

set.all.ident 	 now a synonym for SetAllIdent

rename.ident 	 now a synonym for RenameIdent

posterior.plot 	 now a synonym for PosteriorPlot

map.cell 	 now a synonym for MapCell

get.centroids 	 now a synonym for GetCentroids

refined.mapping 	 now a synonym for RefinedMapping

initial.mapping 	 now a synonym for InitialMapping

calc.insitu 	 now a synonym for CalcInsitu

fit.gene.k 	 now a synonym for FitGeneK

fit.gene.mix 	 now a synonym for FitGeneMix

addSmoothedScore 	 now a synonym for AddSmoothedScore

addImputedScore 	 now a synonym for AddImputedScore

getNewScore 	 now a synonym for GetNewScore

calcNoiseModels 	 now a synonym for CalcNoiseModels

feature.plot.keynote 	 now a synonym for FeaturePlotKeynote

feature.heatmap 	 now a synonym for FeatureHeatmap

ica.plot 	 now a synonym for ICAPlot

dim.plot 	 now a synonym for DimPlot

spatial.de 	 now a synonym for SpatialDe

DBclust_dimension 	 now a synonym for DBClustDimension

Kclust_dimension 	 now a synonym for KClustDimension

pca.sig.genes 	 now a synonym for PCASigGenes

doHeatMap 	 now a synonym for DoHeatMap

icHeatmap 	 now a synonym for ICHeatmap

doKMeans 	 now a synonym for DoKMeans

genes.in.cluster 	 now a synonym for GenesInCluster

kMeansHeatmap 	 now a synonym for KMeansHeatmap

cell.cor.matrix 	 now a synonym for CellCorMatrix

gene.cor.matrix 	 now a synonym for GeneCorMatrix

calinskiPlot 	 now a synonym for CalinskiPlot

dot.plot 	 now a synonym for DotPlot

addMetaData 	 now a synonym for AddMetaData

removePC 	 now a synonym for RemovePC

geneScorePlot 	 now a synonym for GeneScorePlot

cellPlot 	 now a synonym for CellPlot

jackStraw.permutation.test 	 now a synonym for JackStrawPermutationTest

jackStrawMC 	 now a synonym for JackStrawMC

jackStrawFull 	 now a synonym for JackStrawFull

[Package Seurat version 1.3.5 Index]