TSNEPlot {Seurat}	R Documentation

Plot tSNE map

Description

Graphs the output of a tSNE analysis Cells are colored by their identity class.

Usage

TSNEPlot(object, do.label = FALSE, pt.size = 1, label.size = 4,
  cells.use = NULL, colors.use = NULL, ...)

Arguments

object
Seurat object

do.label
FALSE by default. If TRUE, plots an alternate view where the center of each cluster is labeled

pt.size
Set the point size

label.size
Set the size of the text labels

cells.use
Vector of cell names to use in the plot.

colors.use
Manually set the color palette to use for the points

...
Additional parameters to DimPlot, for example, which dimensions to plot.

Details

This function is a wrapper for DimPlot. See ?DimPlot for a full list of possible arguments which can be passed in here.

See Also

DimPlot

[Package Seurat version 1.3.5 Index]