ValidateClusters {Seurat}	R Documentation

Cluster Validation

Description

Methods for validating the legitimacy of clusters using classification. SVMs are used as the basis for the classification. Merging is done based on the connectivity from an SNN graph.

Usage

ValidateClusters(object, pc.use = NULL, top.genes = 30,
  min.connectivity = 0.01, acc.cutoff = 0.9, verbose = TRUE)

Arguments

object
Seurat object

pc.use
Which PCs to use for model construction

top.genes
Use the top X genes for model construction

min.connectivity
Threshold of connectedness for comparison of two clusters

acc.cutoff
Accuracy cutoff for classifier

verbose
Controls whether to display progress and merge results

Value

Returns a Seurat object, object@ident has been updated with new cluster info

[Package Seurat version 1.3.5 Index]