ValidateSpecificClusters {Seurat}	R Documentation

Specific Cluster Validation

Description

Methods for validating the legitimacy of two specific clusters using classification. SVMs are used as the basis for the classification. Merging is done based on the connectivity from an SNN graph.

Usage

ValidateSpecificClusters(object, cluster1 = NULL, cluster2 = 1,
  pc.use = 2, top.genes = 30, acc.cutoff = 0.9)

Arguments

object
Seurat object

cluster1
First cluster to check classification

cluster2
Second cluster to check with classification

pc.use
Which PCs to use for model construction

top.genes
Use the top X genes for model construction

acc.cutoff
Accuracy cutoff for classifier

Value

Returns a Seurat object, object@ident has been updated with new cluster info

[Package Seurat version 1.3.5 Index]