WhichCells {Seurat}	R Documentation

Identify matching cells

Description

Returns a list of cells that match a particular query (usually, query is based on identity class). For example, to find the names of all cells in cluster 1.

Usage

WhichCells(object, value = 1, id = NULL)

Arguments

object
Seurat object

value
Query value to match

id
Variable to query (by default, identity class)

Value

A vector of cell names

[Package Seurat version 1.3.5 Index]