AddImputedScore {Seurat}	R Documentation

Calculate imputed expression values

Description

Uses L1-constrained linear models (LASSO) to impute single cell gene expression values.

Usage

AddImputedScore(object, genes.use = NULL, genes.fit = NULL, s.use = 20,
  do.print = FALSE, gram = TRUE)

Arguments

object
Seurat object

genes.use
A vector of genes (predictors) that can be used for building the LASSO models.

genes.fit
A vector of genes to impute values for

s.use
Maximum number of steps taken by the algorithm (lower values indicate a greater degree of smoothing)

do.print
Print progress (output the name of each gene after it has been imputed).

gram
The use.gram argument passed to lars

Value

Returns a Seurat object where the imputed values have been added to object@data

[Package Seurat version 1.3.5 Index]