BuildClusterTree {Seurat}	R Documentation

Phylogenetic Analysis of Identity Classes

Description

Constructs a phylogenetic tree relating the 'average' cell from each identity class. Tree is estimated based on a distance matrix constructed in either gene expression space or PCA space.

Usage

BuildClusterTree(object, genes.use = NULL, pcs.use = NULL, SNN.use = NULL,
  do.plot = TRUE, do.reorder = FALSE, reorder.numeric = FALSE)

Arguments

object
Seurat object

genes.use
Genes to use for the analysis. Default is the set of variable genes (object@var.genes). Assumes pcs.use=NULL (tree calculated in gene expression space)

pcs.use
If set, tree is calculated in PCA space, using the eigenvalue-weighted euclidean distance across these PC scores.

SNN.use
If SNN is passed, build tree based on SNN graph connectivity between clusters

do.plot
Plot the resulting phylogenetic tree

do.reorder
Re-order identity classes (factor ordering), according to position on the tree. This groups similar classes together which can be helpful, for example, when drawing violin plots.

reorder.numeric
Re-order identity classes according to position on the tree, assigning a numeric value ('1' is the leftmost node)

Details

Note that the tree is calculated for an 'average' cell, so gene expression or PC scores are averaged across all cells in an identity class before the tree is constructed.

Value

A Seurat object where the cluster tree is stored in object@cluster.tree[[1]]

[Package Seurat version 1.3.5 Index]