CellPlot {Seurat}	R Documentation

Cell-cell scatter plot

Description

Creates a plot of scatter plot of genes across two single cells

Usage

CellPlot(object, cell1, cell2, gene.ids = NULL, col.use = "black",
  nrpoints.use = Inf, pch.use = 16, cex.use = 0.5, do.ident = FALSE,
  ...)

Arguments

object
Seurat object

cell1
Cell 1 name (can also be a number, representing the position in object@cell.names)

cell2
Cell 2 name (can also be a number, representing the position in object@cell.names)

gene.ids
Genes to plot (default, all genes)

col.use
Colors to use for the points

nrpoints.use
Parameter for smoothScatter

pch.use
Point symbol to use

cex.use
Point size

do.ident
FALSE by default. If TRUE, allows you to click on individual points to reveal gene names (hit ESC to stop)

...
Additional arguments to pass to smoothScatter

Value

No return value (plots a scatter plot)

[Package Seurat version 1.3.5 Index]