DoHeatmap {Seurat}	R Documentation

Gene expression heatmap

Description

Draws a heatmap of single cell gene expression using the heatmap.2 function.

Usage

DoHeatmap(object, cells.use = NULL, genes.use = NULL, disp.min = -2.5,
  disp.max = 2.5, draw.line = TRUE, do.return = FALSE,
  order.by.ident = TRUE, col.use = pyCols, slim.col.label = FALSE,
  group.by = NULL, remove.key = FALSE, cex.col = NULL, do.scale = TRUE,
  ...)

Arguments

object
Seurat object

cells.use
Cells to include in the heatmap (default is all cells)

genes.use
Genes to include in the heatmap (ordered)

disp.min
Minimum display value (all values below are clipped)

disp.max
Maximum display value (all values above are clipped)

draw.line
Draw vertical lines delineating cells in different identity classes.

do.return
Default is FALSE. If TRUE, return a matrix of scaled values which would be passed to heatmap.2

order.by.ident
Order cells in the heatmap by identity class (default is TRUE). If FALSE, cells are ordered based on their order in cells.use

col.use
Color palette to use

slim.col.label
if (order.by.ident==TRUE) then instead of displaying every cell name on the heatmap, display only the identity class name once for each group

group.by
If (order.by.ident==TRUE) default, you can group cells in different ways (for example, orig.ident)

remove.key
Removes the color key from the plot.

cex.col
positive numbers, used as cex.axis in for the column axis labeling. The defaults currently only use number of columns

do.scale
whether to use the data or scaled data

...
Additional parameters to heatmap.2. Common examples are cexRow and cexCol, which set row and column text sizes

Value

If do.return==TRUE, a matrix of scaled values which would be passed to heatmap.2. Otherwise, no return value, only a graphical output

[Package Seurat version 1.3.5 Index]