ICTopGenes {Seurat}	R Documentation

Find genes with highest ICA scores

Description

Return a list of genes with the strongest contribution to a set of indepdendent components

Usage

ICTopGenes(object, ic.use = 1, num.genes = 30, do.balanced = FALSE)

Arguments

object
Seurat object

ic.use
Independent components to use

num.genes
Number of genes to return

do.balanced
Return an equal number of genes with both + and - IC scores.

Value

Returns a vector of genes

[Package Seurat version 1.3.5 Index]