JackStraw {Seurat}	R Documentation

Determine statistical significance of PCA scores.

Description

Randomly permutes a subset of data, and calculates projected PCA scores for these 'random' genes. Then compares the PCA scores for the 'random' genes with the observed PCA scores to determine statistical signifance. End result is a p-value for each gene's association with each principal component.

Usage

JackStraw(object, num.pc = 30, num.replicate = 100, prop.freq = 0.01,
  do.print = FALSE, rev.pca = FALSE)

Arguments

object
Seurat object

num.pc
Number of PCs to compute significance for

num.replicate
Number of replicate samplings to perform

prop.freq
Proportion of the data to randomly permute for each replicate

do.print
Print the number of replicates that have been processed.

rev.pca
By default computes the PCA on the cell x gene matrix. Setting to true will compute it on gene x cell matrix. This should match what was set when the intial PCA was run.

Value

Returns a Seurat object where object@jackStraw.empP represents p-values for each gene in the PCA analysis. If ProjectPCA is subsequently run, object@jackStraw.empP.full then represents p-values for all genes.

References

Inspired by Chung et al, Bioinformatics (2014)

[Package Seurat version 1.3.5 Index]