JackStrawPlot {Seurat}	R Documentation

JackStraw Plot

Description

Plots the results of the JackStraw analysis for PCA significance. For each PC, plots a QQ-plot comparing the distribution of p-values for all genes across each PC, compared with a uniform distribution. Also determines a p-value for the overall significance of each PC (see Details).

Usage

JackStrawPlot(object, PCs = 1:5, nCol = 3, score.thresh = 1e-05,
  plot.x.lim = 0.1, plot.y.lim = 0.3)

Arguments

object
Seurat plot

PCs
Which PCs to examine

nCol
Number of columns

score.thresh
Threshold to use for the proportion test of PC significance (see Details)

plot.x.lim
X-axis maximum on each QQ plot.

plot.y.lim
Y-axis maximum on each QQ plot.

Details

Significant PCs should show a p-value distribution (black curve) that is strongly skewed to the left compared to the null distribution (dashed line) The p-value for each PC is based on a proportion test comparing the number of genes with a p-value below a particular threshold (score.thresh), compared with the proportion of genes expected under a uniform distribution of p-values.

Value

No value returned, just the PC plots.

Author(s)

Thanks to Omri Wurtzel for integrating with ggplot

[Package Seurat version 1.3.5 Index]