PCHeatmap {Seurat}	R Documentation

Principal component heatmap

Description

Draws a heatmap focusing on a principal component. Both cells and genes are sorted by their principal component scores. Allows for nice visualization of sources of heterogeneity in the dataset.

Usage

PCHeatmap(object, pc.use = 1, cells.use = NULL, num.genes = 30,
  use.full = FALSE, disp.min = -2.5, disp.max = 2.5, do.return = FALSE,
  col.use = pyCols, use.scale = TRUE, do.balanced = FALSE,
  remove.key = FALSE, label.columns = NULL, ...)

Arguments

object
Seurat object

pc.use
Principal components to use

cells.use
A list of cells to plot. If numeric, just plots the top cells.

num.genes
Number of genes to return

use.full
Use the full PCA (projected PCA). Default i s FALSE

disp.min
Minimum display value (all values below are clipped)

disp.max
Maximum display value (all values above are clipped)

do.return
Default is FALSE. If TRUE, return a matrix of scaled values which would be passed to heatmap.2

col.use
Color palette to use

use.scale
Default is TRUE: plot scaled data. If FALSE, plot raw data on the heatmap.

do.balanced
Return an equal number of genes with both + and - PC scores.

remove.key
Removes the color key from the plot.

label.columns
Whether to label the columns. Default is TRUE for 1 PC, FALSE for > 1 PC

...
Additional parameters to heatmap.2. Common examples are cexRow and cexCol, which set row and column text sizes

Value

If do.return==TRUE, a matrix of scaled values which would be passed to heatmap.2. Otherwise, no return value, only a graphical output

[Package Seurat version 1.3.5 Index]