Setup {Seurat}	R Documentation

Setup Seurat object

Description

Setup and initialize basic parameters of the Seurat object

Usage

Setup(object, project, min.cells = 3, min.genes = 1000, is.expr = 0,
  do.logNormalize = T, total.expr = 10000, do.scale = TRUE,
  do.center = TRUE, names.field = 1, names.delim = "_",
  meta.data = NULL, save.raw = TRUE)

Arguments

object
Seurat object

project
Project name (string)

min.cells
Include genes with detected expression in at least this many cells

min.genes
Include cells where at least this many genes are detected

is.expr
Expression threshold for 'detected' gene

do.logNormalize
whether to normalize the expression data per cell and transform to log space.

total.expr
scale factor in the log normalization

do.scale
In object@scale.data, perform row-scaling (gene-based z-score)

do.center
In object@scale.data, perform row-centering (gene-based centering)

names.field
For the initial identity class for each cell, choose this field from the cell's column name

names.delim
For the initial identity class for each cell, choose this delimiter from the cell's column name

meta.data
Additional metadata to add to the Seurat object. Should be a data frame where the rows are cell names, and the columns are additional metadata fields

save.raw
TRUE by default. If FALSE, do not save the unmodified data in object@raw.data which will save memory downstream for large datasets

Value

Seurat object. Fields modified include object@data, object@scale.data, object@data.info, object@ident

[Package Seurat version 1.3.5 Index]