seurat {Seurat}	R Documentation

The Seurat Class

Description

The Seurat object is the center of each single cell analysis. It stores all information associated with the dataset, including data, annotations, analyes, etc. All that is needed to construct a Seurat object is an expression matrix (rows are genes, columns are cells), which should be log-scale

Details

Each Seurat object has a number of slots which store information. Key slots to access are listed below.

Slots

raw.data:

"ANY", The raw project data

data:

"ANY", The expression matrix (log-scale)

scale.data:

"ANY", The scaled (after z-scoring each gene) expression matrix. Used for PCA, ICA, and heatmap plotting

var.genes:

"vector", Variable genes across single cells

is.expr:

"numeric", Expression threshold to determine if a gene is expressed

ident:

"vector", The 'identity class' for each single cell

data.info:

"data.frame", Contains information about each cell, starting with # of genes detected (nGene) the original identity class (orig.ident), user-provided information (through AddMetaData), etc.

project.name:

"character", Name of the project (for record keeping)

pca.x:

"data.frame", Gene projection scores for the PCA analysis

pca.x.full:

"data.frame", Gene projection scores for the projected PCA (contains all genes)

pca.rot:

"data.frame", The rotation matrix (eigenvectors) of the PCA

ica.x:

"data.frame", Gene projection scores for ICA

ica.rot:

"data.frame", The estimated source matrix from ICA

tsne.rot:

"data.frame", Cell coordinates on the t-SNE map

mean.var:

"data.frame", The output of the mean/variability analysis for all genes

imputed:

"data.frame", Matrix of imputed gene scores

final.prob:

"data.frame", For spatial inference, posterior probability of each cell mapping to each bin

insitu.matrix:

"data.frame", For spatial inference, the discretized spatial reference map

cell.names:

"vector", Names of all single cells (column names of the expression matrix)

cluster.tree:

"list", List where the first element is a phylo object containing the phylogenetic tree relating different identity classes

snn.sparse:

"dgCMatrix", Sparse matrix object representation of the SNN graph

snn.dense:

"matrix", Dense matrix object representation of the SNN graph

snn.k:

"numeric", k used in the construction of the SNN graph

[Package Seurat version 1.3.5 Index]