SubsetData {Seurat}	R Documentation

Return a subset of the Seurat object

Description

Creates a Seurat object containing only a subset of the cells in the original object. Takes either a list of cells to use as a subset, or a parameter (for example, a gene), to subset on.

Usage

SubsetData(object, cells.use = NULL, subset.name = NULL, ident.use = NULL,
  accept.low = -Inf, accept.high = Inf, do.center = F, do.scale = F,
  max.cells.per.ident = Inf, ...)

Arguments

object
Seurat object

cells.use
A vector of cell names to use as a subset. If NULL (default), then this list will be computed based on the next three arguments. Otherwise, will return an object consissting only of these cells

subset.name
Parameter to subset on. Eg, the name of a gene, PC1, a column name in object@data.info, etc. Any argument that can be retreived using FetchData

ident.use
Create a cell subset based on the provided identity classes

accept.low
Low cutoff for the parameter (default is -Inf)

accept.high
High cutoff for the parameter (default is Inf)

do.center
Recenter the new object@scale.data

do.scale
Rescale the new object@scale.data. FALSE by default

max.cells.per.ident
Can be used to downsample the data to a certain max per cell ident. Default is inf.

...
Additional arguments to be passed to FetchData (for example, use.imputed=TRUE)

Value

Returns a Seurat object containing only the relevant subset of cells

[Package Seurat version 1.3.5 Index]