VlnPlot {Seurat}	R Documentation

Single cell violin plot

Description

Draws a violin plot of single cell data (gene expression, metrics, PC scores, etc.)

Usage

VlnPlot(object, features.plot, nCol = NULL, ylab.max = 12, do.ret = TRUE,
  do.sort = FALSE, size.x.use = 16, size.y.use = 16,
  size.title.use = 20, adjust.use = 1, size.use = 1, cols.use = NULL,
  group.by = "ident", y.log = F, ...)

Arguments

object
Seurat object

features.plot
Features to plot (gene expression, metrics, PC scores, anything that can be retreived by FetchData)

nCol
Number of columns if multiple plots are displayed

ylab.max
Maximum ylab value

do.ret
FALSE by default. If TRUE, returns a list of ggplot objects.

do.sort
Sort identity classes (on the x-axis) by the average expression of the attribute being potted

size.x.use
X axis font size

size.y.use
Y axis font size

size.title.use
Title font size

adjust.use
Adjust parameter for geom_violin

size.use
Point size for geom_violin

cols.use
Colors to use for plotting

group.by
Group (color) cells in different ways (for example, orig.ident)

y.log
plot T axis on log scale

...
additional parameters to pass to FetchData (for example, use.imputed, use.scaled, use.raw)

Value

By default, no return, only graphical output. If do.return=TRUE, returns a list of ggplot objects.

[Package Seurat version 1.3.5 Index]